requirements.txt
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
docopt==0.6.2
idna==3.4
IPTCInfo3==2.1.4
markdown-it-py==2.2.0
mdurl==0.1.2
numpy==1.24.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
paramiko==3.1.0
Pillow==9.5.0
pycparser==2.21
Pygments==2.15.1
PyNaCl==1.5.0
python-dotenv==1.0.0
requests==2.30.0
rich==13.3.5
shellingham==1.5.0.post1
typer==0.9.0
typing_extensions==4.5.0
urllib3==2.0.2
yarg==0.1.9
openvino==2023.2.0
openvino-telemetry==2023.2.1